+++ build-tmp/source/OriginalFractal.java
@@ -22,7 +22,7 @@ public void setup()
 public void draw() 
 {
   background(0);
-  drawCircle(width/2, height/2, 250);
+  drawCircle((int)width/2, (int)height/2, 250);
 }
 
 public void drawCircle(int x, int y, int radius) 
@@ -32,13 +32,13 @@ public void drawCircle(int x, int y, int radius)
   if (radius > 10) 
   {
     stroke(229, 162, 250);
-    drawCircle(x + radius/2, y, radius/2);
+    drawCircle(x + (int)radius/2, y, (int)radius/2);
     stroke(137, 18, 131);
-    drawCircle(x - radius/2, y, radius/2);
+    drawCircle(x - (int)radius/2, y, (int)radius/2);
     stroke(162, 237, 250);
-    drawCircle(x, y + radius/2, radius/2);
+    drawCircle(x, y + (int)radius/2, (int)radius/2);
     stroke(84, 70, 255);
-    drawCircle(x, y - radius/2, radius/2);
+    drawCircle(x, y - (int)radius/2, (int)radius/2);
   }
 }
   static public void main(String[] passedArgs) {
